tripouille all done, need fix on unit test

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: arthur <arthur@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/09/07 15:35:19 by apigeon           #+#    #+#             */
-/*   Updated: 2021/11/22 14:45:25 by arthur           ###   ########.fr       */
+/*   Updated: 2021/11/27 14:32:22 by apigeon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *s)
+#include "libft.h"
+
+size_t	ft_strlen(const char *s)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i])
